repos/current-builds/knowledge-palace/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-palace"
version = "0.1.0"
description = "Repo-native, provenance-first personal knowledge operating system."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "jsonschema>=4.23.0,<5",
  "PyYAML>=6.0.2,<7",
]

[project.scripts]
knowledge-palace = "knowledge_palace.cli:main"

[tool.setuptools.packages.find]
include = ["knowledge_palace*"]
